diagnose_numeric

Inputs a dataframe and returns various summary statistics of the numeric columns. For example zeros returns the ratio of 0 values in that column. minus counts negative values and infs counts Inf values. Other rarer metrics are also returned that may be helpful for quick diagnosis or understanding of numeric data. mode returns the most common value in the column (chooses at random in case of tie) , and mode_ratio returns its frequency as a ratio of the total rows

diagnose_numeric(.data, ...)

Arguments

.data

dataframe

...

tidyselect. Default: all numeric columns

Value

dataframe

Examples

iris %>%
diagnose_numeric() %>%
print(width = Inf)
#> 150 rows
#> # A tibble: 4 × 11
#>   variables    zeros minus  infs   min  mean   max `|x|<=1 (ratio)`
#>   <chr>        <int> <int> <int> <dbl> <dbl> <dbl>            <dbl>
#> 1 Sepal.Length     0     0     0   4.3  5.84   7.9          0      
#> 2 Sepal.Width      0     0     0   2    3.06   4.4          0      
#> 3 Petal.Length     0     0     0   1    3.76   6.9          0.00667
#> 4 Petal.Width      0     0     0   0.1  1.20   2.5          0.38   
#>   integer_ratio  mode mode_ratio
#>           <dbl> <dbl>      <dbl>
#> 1        0.113    5       0.0667
#> 2        0.187    3       0.173 
#> 3        0.0867   1.4     0.0867
#> 4        0.0867   0.2     0.193