view_missing

View rows of the dataframe where columns in the tidyselect specification contain missings by default, detects missings in any column. The result is by default displayed in the viewer pane. Can be returned as a tibble optionally.

view_missing(df, ..., view = TRUE)

Arguments

df

dataframe

...

tidyselect

view

logical. if false, returns tibble

Value

tibble

Examples

iris %>%
framecleaner::make_na(Species, vec = "setosa") %>%
view_missing(view = FALSE)
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1           5.1         3.5          1.4         0.2    <NA>
#> 2           4.9         3.0          1.4         0.2    <NA>
#> 3           4.7         3.2          1.3         0.2    <NA>
#> 4           4.6         3.1          1.5         0.2    <NA>
#> 5           5.0         3.6          1.4         0.2    <NA>
#> 6           5.4         3.9          1.7         0.4    <NA>
#> 7           4.6         3.4          1.4         0.3    <NA>
#> 8           5.0         3.4          1.5         0.2    <NA>
#> 9           4.4         2.9          1.4         0.2    <NA>
#> 10          4.9         3.1          1.5         0.1    <NA>
#> 11          5.4         3.7          1.5         0.2    <NA>
#> 12          4.8         3.4          1.6         0.2    <NA>
#> 13          4.8         3.0          1.4         0.1    <NA>
#> 14          4.3         3.0          1.1         0.1    <NA>
#> 15          5.8         4.0          1.2         0.2    <NA>
#> 16          5.7         4.4          1.5         0.4    <NA>
#> 17          5.4         3.9          1.3         0.4    <NA>
#> 18          5.1         3.5          1.4         0.3    <NA>
#> 19          5.7         3.8          1.7         0.3    <NA>
#> 20          5.1         3.8          1.5         0.3    <NA>
#> 21          5.4         3.4          1.7         0.2    <NA>
#> 22          5.1         3.7          1.5         0.4    <NA>
#> 23          4.6         3.6          1.0         0.2    <NA>
#> 24          5.1         3.3          1.7         0.5    <NA>
#> 25          4.8         3.4          1.9         0.2    <NA>
#> 26          5.0         3.0          1.6         0.2    <NA>
#> 27          5.0         3.4          1.6         0.4    <NA>
#> 28          5.2         3.5          1.5         0.2    <NA>
#> 29          5.2         3.4          1.4         0.2    <NA>
#> 30          4.7         3.2          1.6         0.2    <NA>
#> 31          4.8         3.1          1.6         0.2    <NA>
#> 32          5.4         3.4          1.5         0.4    <NA>
#> 33          5.2         4.1          1.5         0.1    <NA>
#> 34          5.5         4.2          1.4         0.2    <NA>
#> 35          4.9         3.1          1.5         0.2    <NA>
#> 36          5.0         3.2          1.2         0.2    <NA>
#> 37          5.5         3.5          1.3         0.2    <NA>
#> 38          4.9         3.6          1.4         0.1    <NA>
#> 39          4.4         3.0          1.3         0.2    <NA>
#> 40          5.1         3.4          1.5         0.2    <NA>
#> 41          5.0         3.5          1.3         0.3    <NA>
#> 42          4.5         2.3          1.3         0.3    <NA>
#> 43          4.4         3.2          1.3         0.2    <NA>
#> 44          5.0         3.5          1.6         0.6    <NA>
#> 45          5.1         3.8          1.9         0.4    <NA>
#> 46          4.8         3.0          1.4         0.3    <NA>
#> 47          5.1         3.8          1.6         0.2    <NA>
#> 48          4.6         3.2          1.4         0.2    <NA>
#> 49          5.3         3.7          1.5         0.2    <NA>
#> 50          5.0         3.3          1.4         0.2    <NA>